Move resolveMBQLField to separate module
